+++ Changes.md
@@ -23,7 +23,7 @@ API
 ---
 
 - LocalDispatcher :
-  - Added `Job.status()` method.
+  - Added `Job.status()` and `Job.statusChangedSignal()` methods.
   - Added `JobPool.addJob()` and `JobPool.removeJob()` methods.
 - GafferTractor : Added `tractorAPI()` method used for accessing the `tractor.api.author` module.
 - GafferTractorTest : Added `tractorAPI()` method which returns a mock API if Tractor is not available. This allows the GafferTractor module to be tested without Tractor being installed.